requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==4.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
django-mptt==0.13.4
django-redis==5.0.0
django-widget-tweaks==1.4.8
idna==3.3
mysqlclient==2.0.3
Pillow==8.3.2
pycodestyle==2.7.0
PyMySQL==1.0.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
soupsieve==2.3.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
